Delete legacy RT archiver (#2110)
* remove gtfsrt-v3 node pool

* delete legacy rt k8s configurations

* delete gtfs-rt-archive application code; keep 2 functions for agencies_verify

* remove legacy archiver k8s docs

* delete more docs

* indicate deprecated in docs

* update docs

* maybe fix docs build
