--- ci/vars/releases/preprod-gtfs-rt-archive.env
@@ -1,2 +1,0 @@
-RELEASE_DRIVER=kustomize
-RELEASE_KUSTOMIZE_DIR=kubernetes/apps/overlays/gtfs-rt-archive-preprod
