--- kubernetes/apps/manifests/gtfs-rt-archive/app.yaml
@@ -1,59 +1,0 @@
----
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: gtfs-rt-archive
-  labels:
-    name: gtfs-rt-archive
-spec:
-  replicas: 1
-  strategy:
-    type: Recreate
-  selector:
-    matchLabels:
-      name: gtfs-rt-archive
-  template:
-    metadata:
-      labels:
-        name: gtfs-rt-archive
-    spec:
-      containers:
-        - name: app
-          image: gtfs-rt-archive:latest
-          envFrom:
-            - configMapRef:
-                name: archiver-app-vars
-            - configMapRef:
-                name: archiver-channel-vars
-          volumeMounts:
-            - name: agencies-data
-              mountPath: /secrets/agencies-data
-            - name: gcs-upload-svcacct
-              mountPath: /secrets/gcs-upload-svcacct
-          resources:
-            requests:
-              memory: 512Mi
-              cpu: 5.0
-            limits:
-              memory: 1Gi
-      volumes:
-        - name: agencies-data
-          secret:
-            secretName: agencies-data
-        - name: gcs-upload-svcacct
-          secret:
-            secretName: gcs-upload-svcacct
-      tolerations:
-        - key: resource-domain
-          operator: Equal
-          value: gtfsrt
-          effect: NoSchedule
-      affinity:
-        nodeAffinity:
-          requiredDuringSchedulingIgnoredDuringExecution:
-            nodeSelectorTerms:
-            - matchExpressions:
-              - key: resource-domain
-                operator: In
-                values:
-                - gtfsrt
