--- kubernetes/apps/manifests/gtfs-rt-archive/kustomization.yaml
@@ -1,3 +1,0 @@
-resources:
-- app.yaml
-- archiver-app-vars.yaml
