--- kubernetes/apps/overlays/gtfs-rt-archive-preprod/image-pull-policy.patch.yaml
@@ -1,11 +1,0 @@
----
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: gtfs-rt-archive
-spec:
-  template:
-    spec:
-      containers:
-        - name: app
-          imagePullPolicy: Always
