--- kubernetes/apps/overlays/gtfs-rt-archive-preprod/kustomization.yaml
@@ -1,11 +1,0 @@
-apiVersion: kustomize.config.k8s.io/v1beta1
-kind: Kustomization
-namespace: gtfs-rt-preprod
-
-resources:
-- ns.yaml
-- archiver-channel-vars.yaml
-- ../gtfs-rt-archive-release
-
-patches:
-- image-pull-policy.patch.yaml
