--- kubernetes/apps/overlays/gtfs-rt-archive-prod/ns.yaml
@@ -1,5 +1,0 @@
----
-apiVersion: v1
-kind: Namespace
-metadata:
-  name: gtfs-rt
