--- kubernetes/apps/overlays/gtfs-rt-archive-release/kustomization.yaml
@@ -1,10 +1,0 @@
-apiVersion: kustomize.config.k8s.io/v1beta1
-kind: Kustomization
-
-resources:
-- '../../manifests/gtfs-rt-archive'
-
-images:
-- name: 'gtfs-rt-archive'
-  newName: 'us.gcr.io/cal-itp-data-infra/gtfs-rt-archive'
-  newTag: '2.1'
