--- services/gtfs-rt-archive/.gitignore
@@ -1,1 +1,0 @@
-.env
