--- services/gtfs-rt-archive/build.env
@@ -1,1 +1,0 @@
-BUILD_REPO=us.gcr.io/cal-itp-data-infra/gtfs-rt-archive
